pyproject.toml
[tool.poetry]
name = "mcuthallclubwebsite"
version = "0.1.0"
description = ""
authors = ["xanmankey <86136010+xanmankey@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
flask-htmx = "^0.3.2"
jinja2-fragments = "^1.6.0"
wtforms = "^3.1.2"
flask-admin = "^1.6.1"
jinjax = "^0.45"
active-alchemy = {git = "https://github.com/xanmankey/active-alchemy.git"}
flask-login = "^0.6.3"
flask-mail = "^0.10.0"
sqlalchemy = "1.4.0"
python-dotenv = "1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
